engine/pyproject.toml
[project]
name = "afls"
version = "0.0.1"
description = "Typed reasoning engine for AI-toward-suffering-reduction"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40",
    "httpx>=0.27",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.9",
    "typer>=0.15",
]

[project.scripts]
afls = "afls.cli:app"

[dependency-groups]
dev = [
    "mypy>=1.13",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "types-pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["afls"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_unimported = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
